.gentle-ai/skills/system-domain-builder/scripts/requirements.txt
PyPDF2>=3.0.0
Pillow>=10.2.0
